General refactoring to make it better readable
