+++ src/test/java/com/amazon/jenkins/ec2fleet/EC2FleetCloudTest.java
@@ -4,7 +4,8 @@ import com.amazon.jenkins.ec2fleet.fleet.AutoScalingGroupFleet;
 import com.amazon.jenkins.ec2fleet.fleet.EC2Fleet;
 import com.amazon.jenkins.ec2fleet.fleet.EC2Fleets;
 import com.amazon.jenkins.ec2fleet.fleet.EC2SpotFleet;
-import com.amazon.jenkins.ec2fleet.fleet.RegionInfo;
+import com.amazon.jenkins.ec2fleet.utils.RegionInfo;
+import com.amazon.jenkins.ec2fleet.utils.AwsPermissionChecker;
 import com.amazonaws.regions.RegionUtils;
 import com.amazonaws.services.ec2.AmazonEC2;
 import com.amazonaws.services.ec2.AmazonEC2Client;
